requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pymongo>=4.5.0
requests>=2.31.0

# Data analysis and metrics
pandas>=2.0.0
numpy>=1.24.0

# Streamlit frontend
streamlit>=1.28.0
plotly>=5.17.0

# Environment variables
python-dotenv>=1.0.0

# Broker SDKs (commented out - install as needed)
# ib-insync>=0.9.86  # Interactive Brokers
# kiteconnect>=4.2.0  # Zerodha
# python-binance>=1.0.19  # Binance
# MetaTrader5>=5.0.45  # For Vantage (MT5)

# Optional: Performance metrics
# quantstats>=0.0.62

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.1.0
